Split up settingsmenu code. Clean up help icons.
